Overhauls install dependencies code
In the install_dep package within the main package is some
code to help install all the dependcies to get full
functionality. All aspects of script have NOT been tested and
not all options are available yet.

These files are mainly kept for my sake to easily install
packages. However, these scripts are likely to fail and
it thus it is advised that the use go through and install
required packages manually to make sure that everything is
working correctly.

v0.0.1a