Moves mediainfo to class
This update moves the mediainfo module to a single class.
Don't know why I didn't do this sooner, but now when the
mediainfo class initializes it parses all information from
the mediainfo CLI to a dictionary in an attribute. There
are then 3 methods, one each for getting audio, video, and
text information