Begins adding commercial removal
This version begins to add comercail removal using comskip and
ffmpeg. Comskip is run on a file and then the edl file will be
used with ffmpeg to split up files into multiple segments; i.e.,
everything that isn't commercials. ffmpeg will then join these
segements together into one file without commercials.

Also changed the limitCPUusage utility so that it checks for
the cpulimit cli on import. If cpulimit is NOT found, then
an exception is raised. Thus it is a good idea to import as:

try:
  from limitCPUusage import limitCPUusage;
except:
  limitCPUusage = None;

Then, if you want to limit cpu usage, you can wrap that
code in an if block:

if limitCPUusage:
  cpuid = limitCPUusage( pid, 75, **kwargs)