Gets comremove functional
The comremove class is not functional. Through various methods in
the class, the comskip CLI is run, the edl output file is used
to extract segments that are NOT commericals using ffmpeg, and
the show segments are joined back together with ffmpeg.

While the comremove class does accept input for a comskip ini
file, it will fall back to using the COMSKIP_INI environment
variable. Information on how to set this variable is in the
README.md file