Adds in_place option to videoconverter
This update adds an in_place keyword argument to the
class initiation of the videoconverter class. What this
does is stops the transcode method from creating a directory
tree to place files in. Example below

When the keyword is set to false, a directory tree such as
 Movies/ OR TV Shows/The Series/Seaons XX
are created in the desired output directory. When using in_place
this directory tree is NOT used, and the file is placed in the
top-level (pending subtitles) of the output directory. If no
output directory was explicitly set, then the input file's
parent directory is used

  To Add
    - no_log_file to suppress the handbrake log file creation