Enables in_place & no_hb_log in DVR Post script
This update forces the in_place and no_hb_log keywords for
the videoconverter class to be set when running the
Plex_DVR_PostProcess script. This causes the output file(s) to
be placed in the same directory as the input file and suppresses
creation of the .log and .err files for stdout and stderr,
respectively, when the HandBrakeCLI program is running.

Also fixed bug that arose from copying code from MKV_Cron_Convert
where I tried to access a videoconverter attribute that did not
exist. This is not fixed and the correct attribute
(i.e., transcode_status) is now check for successful completion
of HandBrakeCLI transcoding