Fixes IMDb ID from TVDb bug
The main fix is that there was a bug in the get_IMDb_ID
function where a missplaced ')' caused a crash. This has
been fixed.

Also removed the import of the test_dependecies.py util, as
the checking for dependencies happens in the __init__.py file
on import and within the other functions/classes that require
optional dependencies.

test_dependencies.py is depricated and will be removed in future
versions