Comskip remove now sub-class of subprocManager
Update adds the subprocManager class to the utils package.
This class allows for easy multiprocessing of system class
by allowing the user to add command lists, such as those input
into subprocess.Popen, to a queue in the class and then running
all the commands in the queue in a parallel way. The class has
CPU limiting built-in (if the cpulimit CLI is installed).

The comremove class is now a subclass of the subprocManager class
as the multiple subprocess calls in the comcut method can be
parallelized; i.e., extracting more than 1 non-commercial segment
at a time.

More code changes need in the comremove class to remove reference
to limitCPUusage, as this now contained in the subprocManager
class