Makes videoconverter subclass of subprocManager
Adds subprocManager as parent class of videoconverter; now
mediainfo and subprocManager are parent classes. For this
to work correctly, add super().__init__() class in both mediainfo
and subprocManager classes.

Also fixed bug in getTMDb_Info that arose from None being input
into the attempts keyword, which overrode the default of 3. Thus,
default of 3 is now set inside the function if attempts evaluates
to False.