Change to mediainfo --Output argument
Starting with version 17.10 of mediainfo, you must use the OLDXML
output format for the parsing to work properly, however, older
version require the XML key. Thus, the mediainfo command is
run with --Output=OLDXML and then we try to parse the output.
If this fails, then we use --Output=XML. Future update I will
fix this to check the version of mediainfo, but quick fix for
now.

Also moved the limitCPUusage utility to a new name to see what
breaks