Updates self.log variables in two classes
In comremove, adds a self.log = statement in the __init__ so
that the logging informaiton uses comskip, instead of
subprocManager.

In mediainfo, moved self.log = statement to __init__ so that will
be correct after call to super. Also added self.cmd = to set up
base command list. This just saves some lines of code in the
__parse_output() method as can now do self.cmd + [self.in_file]