Cleans up imports, removes commented code
Imports in the getMetaData function are cleaned up, with failed
to import function names set to None. Can then check if function
name is valid in if statement and use accordingly.

Old, commented out, code in getTVDb_Info was delete.

Moved placement of the self.in_file = in_file in the mediainfo
command as need the self.cmd attribute defined becuase it is
used in __parse_output, which is run when self.in_file is changed