Adds lock file so only one instance
Adds a lock file in /tmp/ so that only one instance can run at a
time. While the lock file exists, the function sleeps for 1
second. The lock file is delete at exit by registering the
os.remove function, with lock_file as argument, using the atexit
module. Will be testing on RPi