Moves Plex_DVR_PostProcess to own file
Moves the Plex_DVR_PostProcess function to its own file in the
video_utils module so that it can be used more easily for
scripting if desired. The Plex_DVR_PostProcess CLI now just
imports the function and uses it.

atexit register has been updated to check for the existance of
the lock file before trying to delete it