Updates logging/lock file in Plex DVR
Update attempts to fix issue with lock file removal and logging
file opening. When the lock file and/or logging file for the
Plex DVR post-processing were opened, only the user that
created them (i.e., the one running the script) could edit or
remove them. This caused the post processing to fail when trying
to open the log file for writing.