Moves lock file to CLI file
In pervious version, the Plex_DVR_PostProcess function created
the .lock file used to stop other processes from runing. However,
to make the script more robust (i.e., remove the lock file on
error), lock file creation/deletion is moved to the CLI file in
the ./bin directory. The call to the Plex_DVR_PostProcess
function is then placed in a try/except, so that even if the
command fails, the script file will continue and remove the lock
file.

Note that with this change, calls to the PostProcess function
from within python will NOT wait for other instances of the
PostProcesser to finish, they will just run!!!