Fixes log handler setup issue
Information is popped off of the plexFMT log format handler.
If running in python interpreter, this would cause error on
second call to function because key would no longer exist. Thus,
add None return on failuer to find key in dictionary. The
varaibles are then checked to see if None.