Updates logging in getIMDb_ID
Logging has been updated in the getIMDb_ID function so that
more debugging information is printed. After some testing, there
seems to be some issue on the Raspberry Pi with getting the IMDb
ID that does not exist on a Mac. Hopefully the extra debugging
will help pin down the issue.