Fixes no outfile in log and code cleanup
The outfile in the videoconverter.transcode() method was not
being printed in the logger. This is now fixed; forgot the {}
in the format string.

In the mediainfo get_audio_info() and get_video_info() methods,
keys in the returned dictionaries have been changes so that they
match the corresponding HandBrake flag. Note that all flags use
the longer -- prefix. All the key/value pairs and then interated
over to check for the -- prefix in the key. If -- exists in the
key, then the key is append to the HandBrake command list. The
value corresonding to the key is then checked to see if it is NOT
empty. If it is NOT empty, then the value is also appended to the
HandBrake command list. This is done because some flags do not
have values associated with them; i.e., they are just flags.
It is worth clarifying that the value of the key/value pair is
only checked if the key has the -- prefix.

v0.1.16