Plex scanner utils and fix to comremove
Quick fix is to comremove where split files were left if something
went wrong. Those files should now be deleted no matter what.

The plex scanner code is added to get around the 'Error occurred
during post processing' error in Plex DVR. From looking at logs
and reading Plex forums, because the transcoder deletes the
input file; i.e., the file Plex created, even though the post
processing completes normally, Plex throws an error because the
source is gone. The transcoded files still get moved to the
right location, so the error doesn't help or tell when there is
actually an error.

To get around this, the input file is not deleted and a new
process is forked after the transcode that will trigger Plex
to scan library files to find the transcoded file and, pending
keywords, delete the original TS file. Checks are performed to
try to limit library scanning to just the TV season where the
new episode is added.

v0.1.17