Check for quick start and remove -map
In the comremove.comcut method, the time for the start of the
comercial is now checked so see if it is less than one (1)
second. If it is, then no cutting is done.

May change code in future to check through all comercials to
make sure that time from one commercial end to the next
commercial start is not too short.

Also removed the '-map 0' option from ffmpeg as it caused a
failure in a file where the second audio stream was corrupted.
Without the '-map 0' option the second audio stream was simply
ignored, with mediainfo showing the same information for the
original and split files, less the one audio stream