Decoupling of forked process
Added 'decoupling' of forked process as I believe it is getting
eaten when a new Plex_DVR_PostProcess command begins running.
The new method should truly daemonize the library updating and
file removal.

Also added check in plex_utils to see if the LD_LIBRARY_PATH
directory already exists in the LD_LIBRARY_PATH, if it does,
it is NOT added a second time

v0.1.26