New method for plex post-post process
New method of removing original recording file after post-process
script has finished. Plex sees symlinks as just the destination
file (as it should), so we now create symlink with original
recording name pointing to the transcoded file. This is done
after a successful transcode; the original recording is removed
and a symlink, using relative paths so link will work after Plex
moves it, is created with source of transcoded file and
destination of original file path. Note that the plexDVR_Rename
function still creates a hard link to the original recording with
a name that matches the naming convention. This hard linked file
is then used for transcoding; we do NOT want to mess with the
original file incase something breaks.

After the transcode completes (success or failure), we check if
the renamed/hardlinked file exists and delete it if it does;
the .transcode() method will delete it if successfull, but if
fails will stay and then Plex will see two (2) of the same file.
If the no_remove flag is set, then we do NOT remove the original
recording, else, we delete it an create the sym link

At the CLI level, if the transcode is successful AND the
no_remove flag is NOT set, then we run the plexDVR_Cleanup
fuction, which tries to find the moved file in a smart
manner (using informtion from original recording file name
to build file path) to find and delete the symlink created.
If the smart method fails, a brute force approach is used.

v0.1.27