Update for new plex library dir
With the offical Plex release for ARM processors the
LD_LIBRARY_PATH changed a little with all the Plex utilities in
/usr/lib/plexmediaserver (on Debian) and the libraries in ./lib
So, when grabbing just the LD_LIBRARY_PATH as the path, the
scanner was not found. Updated to scale up the directory tree
looking for the command.

Also add log.exception() calls in try/except statements so that
I can debug thing easier.

v0.1.31