Fixes issues with comskip cmd list
Fixes bug where subsequent calls to comremove led to odd comskip
commands due to the base command (i.e., self._comskip) not being
copied. So, in the comskip method of comremove class when did
cmd = self._comskip, got a pointer to the self._comskip attribute
which was appened to leading to incorrect comskip commands.

v0.1.34