Moves daemon creation to Plex Post CLI
A completely separate process is created in the Plex_DVR_PostProcess
CLI instead of just calling the spawnDaemon() function. This is done
so that the lock file remains until the plex scanning is completed.
When in the fully new process, the plexDVR_Scan utility is run to try to
update Plex so that the new, transcoded file is recognized as a second
version of the original recording. The original file will then be removed
and Plex re-updated to realized the original is gone. Any code after this
if block will only run in the remaining child as all other processes are
exited within this block

v0.1.36