Fixes failed to remove file and Plex move issue
Used wrong variable name when trying to remove the orginal recording from
the destination directory and got a 'file not found error'.

Other updates include addition of a 'subfolder' keyword to the
videoconverter class that, when set to False, disables the creation
of subdirectories to store video and subtitle files. This is done because
it seems that Plex does NOT like moving folders that are found in the
.grab/ directory.

Also changes values returned from Plex_DVR_PostProcess fuction to include
the path to the transcoded file. This values is then fed into the
plexDVR_Scan function so as to check for the existance of the transcoded file.
If the transcoded file is NOT found within the directory the original recording
has been moved to, then there was likely an error when Plex tried to move the
file. Since there is no transcoded file, there is no need to run the Plex
Media Scanner command, so we simply return from the function when transcoded
file not found.

Future Ideas:
  - Enable iteration over all libraries with 'TV' in the name to find
    recordings in stead of just the first library

v0.1.41_ffmpeg