Improves crop detection
Crop detection is improved by looking for crop region using a 20
second chunk of video every 5 minutes until the end of the video
is reached. The median value of all the crop regions, if any are
found, is returned as the cropping filter. While slower than the
first method, it should be more accurate.

v0.1.42_ffmpeg