Adds ffmpeg progress util and updates README
A function named `progress` has been added to the ffmpeg_utils.py
file that logs remaining time left in ffmpeg conversion at a given.
interval. The program takes a Popen instance, and optionally and
interval keyword, and reads from a pipe the output of the ffmpeg
command. This function has NOT been implemented anywhere!

Typos in README have been fixed and the format of the file
changed so that each sentence has its own line.

v0.1.44_ffmpeg