Updates season/episode regex in plex utils
The season/episode regex pattern matching has been updated in
the Plex utils to match that of the changes to videoconverter
class in previous update.

As Plex sometimes uses the year for the season of TV shows, the
regex pattern has been changed to look for 2 or more digits after
both the 's' and 'e' characters in the season/episode part of the
file name.

Also moves some variables to be 'private' variables.

v0.1.46_ffmpeg