Adds number of updates option to ffmpeg progress
The `progress` function in the ffmpeg_utils.py file has been
updated to include the nintervals keyword. This keyword allows
one to specifiy the number of updates they would like the function
to produce during ffmpeg conversions.

When using this keyword, the value of the interval keyword is used
before the first log is printed; i.e., from the time the progess()
function is called to the time the first log is printed is interval.
After this point, the interval is updated based on the remaining
conversion time and the number of logs requested. Note that the
number of logs specified likely has a +-1 range on it depending
on how consistent the conversion speed of the file is.

Note that the progess() function has NOT be implemented in the
videoconverter class yet.

v0.1.47_ffmpeg