Bug fixes, ffmpeg progess enabled, more methods!
The main change is the implementation of the progress() function
from ffmpeg_utils.py. The function is used to log the estimated
completion time of the ffmpeg conversion; however, this is only
enabled when ffmpeg log files are disabled, i.e., the no_ffmpeg_log
keyword is set to True in the videoconverter instance. This only
occurs when logging is disabled because one would be able to look
at the raw output from ffmpeg with logs enabled.

To get the progress() function working, the applyFunc() method has
been added to the subprocManager class. This method takes a function
reference as input, and possibly a tuple or list of argumenst and
a dictionary of keywords/values. The function MUST accept a Popen
instance as the first positional input and is run ONLY if there is
one (1) process currently running; a check is performed to ensure
that only one (1) process is running, errors logged otherwise.

Bug fixes:
  In the mediainfo class, the parsing of the StreamOrder information
  from mediainfo CLI was causing some issues. Was spliting the
  StreamOrder on hyphen (-), then joining on colon (:) to set the
  mapping for ffmpeg as the values for StreamOrder were 'n-n' when
  testing move to ffmpeg. However, it seems that they can be, or have
  switiched to, integer values. Thus, added a try and except where
  try is to split the StreamOrder and except is to assume input
  file of '0' and convert convert StreamOrder to string, giving
  a list of ['0', str('StreamOrder')]. After the try/except,
  the list is joined using colon (:) as before

v0.1.48_ffmpeg