Cleans up code, starts watchdog code additions
Code in the main .transcode() method of the videoconverter
class has been cleaned up a bit.
Also adds creations of a .filename.ext.inprogress file
that is touched when transcode begins.
this file is deleted when transcodings finishes cleanly,
however will persist if errors.
Used to inform code about transcode state if transcode was
canceled or failed.

Adds some signal processing (i.e., SIGINT and SIGTERM) to
the base subprocManager class so that processes are terminited
when one of these messages is caught.

v0.2.0