Removes auto audio downmix and bug fixes
In almost all previous version, if an audio track with more than 2
channels was encountered, the track would be downmixed to stereo
AND copied. After all audio tracks were queued for the conversion,
a final check was done to make sure that there were not multiple
stereo audio tracks of the same language created in the output file;
i.e., we don't want both Dolby and DTS tracks downmixed to stereo.
However, this could lead to the removal of commentary tracks as they
may be stereo.

The reasion for automatic downmixing was so Plex would never have
to transcode audio, however, I have run into issues with artifacts
in the downmixed AAC tracks when playing through Plex. As it is
not too CPU intensive to transcode audio, auto downmixing has been
removed and all audio tracks in the source file are simply copied
to the destination.

Bug fixes were minor variable name typos and such.

v0.2.1