Improves SIG* handling
Handling of SIG* signals is now set up in the main __init__.py
file. A package wide _killEvent is set up in the file, with this
event set when a SIG* is caught.

Both the subprocManager class and the ffmpeg_utils.crop_detect()
function check the _killEvent to determine if they should run or
stop running.

Note that the _killEvent is cleared with each call the
videoconverter.transcode(). This is done so that if you cancel
a transcode, future transcodes are not affected; i.e., you don't
have to manual clear the event or reload python. However,
if using this programatically, such as in a watchdog type service,
be sure to check the _killEvent if a transcode fails, as system
may be shutting down, or some other situation.

v0.2.2