Adds options, bug fixes, and testing
Adds the chapters keyword to comremove that simply creates an
ffmpeg metadata file (suffic .chap) for writing chapter markers
for commercial breaks.
This options has been added to Plex DVR using the destructive
keyword; still must be added to CLI.

The videoconverter class has been updated to check or the
existence of a .chap file with the same basename as the input
file. If the .chap file is found, then it is used to
write chapter information into output file during transcode.

In the getIMDb_ID function, has been updated to search by
series (search_movie), and if series with correct year/title
found, get list of episodes. Then iterate over list of
episodes to find matching episode for metadata.

Bug fixes:
  - Fixes missing audio track information in file name
  - Fixes divide by zero bug in ffmpeg_utils.progress
  - Fixes self.chapterFile bug
  - Fixes returncode index issue that could arise when SIG* handled
  - re_test case (checks for sXXeXX pattern in file name) has been
    removed as this was left over code from <= v0.1.44_ffmpeg

v0.2.4