Begins creating Plex DVR watchdog
As there has been a bug in the Plex DVR PostProcessing feature
for almost a year (if there is more than one file in the
.grab folder, only one (1) file will be moved to final location),
I have decided to create a watch dog that will watch user specified
Plex library directories for files to appear in .grab folder.

When a file appears in .grab, it is logged and when file created
that is NOT in .grab, we check the file against the list of
.grab files. If file moved from .grab, then convert it.

This watchdog program should be started in a service (on Linux).

v0.2.6