Adds service template, refactors code
In the new systemd folder, there is an example
plex_dvr_watchdog.service file for setting up a Linux service
for the plex dvr watchdog.

Code refactoring includes moving IMDb_ID metadata interogation
to a new _metaData method in the videoconverter clas; was in
the file_info method. This should allow for better error handling
in the future.

Changes:
  Logs for the watchdog are now in:
    $HOME/Library/Application Support/video_utils

To do:
  - Add a --no-rename flag to all pertinent functions/classes to
disable renaming files to my convention; this will allow people
to keep the Plex naming convention

v0.2.11