Fixes restart bug and improves signal handling
There was a bug in the converting queue of the Plex DVR watchdog
where a file would be removed from the list of files being converted
even if the conversion failed. If this was the only file being
converted, this would cause the file that tracks the list of
files to convert to be removed. Thus, the file would never be
converted.

Added more checks for the _killEvent being set in the
videoconverter class so that transcode will not start if
event is set before transcode begins.

Needs more testing!!!

v0.2.14