Fixes link creation error, updates signal handling
There was an issue in the Plex utilities when trying to make a
hard link to the original DVR file if the file already exists.
os.link call is now wrapped in try/except and os.rename changed
to os.replace. This should fix this issue.

To improve signal handling for SIGINT and SIGTERM, have changed
the _killEvent that is initialized in main __init__.py file to
_sigintEvent and _sigtermEvent; these events are set when their
respective signals are called. All code has been updated to match
these changes, with the videoconverter class updated to return from
transcode() method right way if the _sigtermEvent is set.
Method then clears the _sigintEvent so that a signal interrupt
does not disable transcoding for all time.

v0.2.15