Fixes queuing bug in Plex DVR watchdog
In the Plex library_watchdog, there was an issue with queuing
recordings to be converted. When a recording was moved from the
.grab folder to its final location, the original location was
removed from the .recordings list and added to the .queue().
However, if the watchdog was stopped at this point, the queued file
would not be processed when the watchdog restarted because it was
never added to the converting list, which cache to file to maintane
list on restart.

To fix this issue, the queue() has been removed and now in the
checkRecording method, if file is recording that has been moved
to final location, it is removed from the recordings list
and added to the converting list. In the processing thread,
instead of trying to get a file from the queue, now try to get
zeroth element of the converting list. If fails, then sleep
and continue while loop. After conversion is done, the file is
removed from the converting list.

v0.2.16