Fixes import bug, plex scan user check
In the MakeMKV_Watchdog, failed to import the plexDVR_Scan
function, which caused an exception when calling the function
in the code.

Added a check of the user running the plexDVR_Scan function
using os.environ['USER']. If user is not PLEX (forced up case),
then return from the function as the Plex Media Scanner must
be run under the plex user, otherwise there are issues.

Cursory testing of MakeMKV_Watchdog seems to be working fine.

Note:
  VobSub2SRT program takes extremely long time to create SRT file.
Not sure what the issue is, but working on python based code
to extract bitmaps from .sub files. Can then run tesseract on
bitmaps for text identification to create srt files.

v0.3.1