Updates series folder to include date; README
When creating the directory path for TV shows, the series directory
will now contain the year; (YYYY); IF it is downloaded in the
metadata. This is done so that directory trees better match Plex.

Main README had been updated to fix some typos. Also removes
information about MKV_Cron_Convert (this is depricated and will
be removed in the future) and adds information about the
MakeMKV_Watchdog.

v0.3.2