Adds MKV tagging; code improvments
Matroska tagging is now possible, using the mkvpropedit CLI
for actual tagging. Metadata are written out to an XML file that
is then used in mkvpropedit to write tags. If/when mutagen fully
supports mkv tagging, will switch to that. The mkvTags function
is responsible for tagging; code is very similar to mp4Tags.

Some changes to the videoconverter class have been made to enable
tagging of mkv files. Not exhaustively tested, so may have
introduceds bugs.

Other changes include:
  - Creation of checkCLI function that check if given CLI is
    installed. This function is now used in all modules
    that depend on CLIs
  - Imporves metadata for the ccextract function
  - Cleans up metadata for videoconverter class

Bug fixes:
  - Fixes bug where .inprogress file persists if transcode
    fails

v0.4.0