Adds user post-process script for plex watchdog
The ability for the user to specify a post-processing script
to run in the Plex_DVR_Watchdog has been added through the
--script flag. Set this flag to the full path to an existing
Plex DVR Post-processing script and that script will be run
when DVRed files are moved to their final location.

The README has been updated to fix typos and reflect changes.

v0.5.0