Changes handling of transcode fail plex
An infinite loop would occur if the ffmpeg transcode failed.
Based on a few days of the watchdog running fine, it seems that
ffmpeg only failed due to a corrupted input file. Thus, the program
now deletes the input file if there is an error; i.e., the original
recording is removed on failed transcode.

Scan of plex directory still takes places; this should remove the
item from the library but needs tested.

v0.6.1