Fixes output directory bug, improves logs
There was a bug where if the out_dir was not set at class
initialization, it would be set to the directory of the
input file on first conversion. On subsequent conversions,
the out_dir would not be changed. This has been fixed and
if the out_dir is not set, then new_out_dir is set to the
directory of the input file. new_out_dir is used in all methods
for setting up output directory.

The in_place flag would also not always work, with Series/Season
info being appended to the input file's directory tree; no longer
in place I would say. This bug has also been fixed.

Logger for the videoconverter and comremove classes have been
moved from .log to .__log so that function name resolution
is accurate and not all methods log to the logger that was
last initialized on class startup.

v0.6.9