Fixes converting file remove on sigterm
In the plex/library_watchdog, would always remove file from
converting list after transcode finished/was stopped. This
is not always what we want to happend; if the program was
interrupted or killed, we want the transcode that was stopped to
start on next run. So, now we check that the _sigintEvent and
_sigtermEvent are NOT set before removing file from converting
list.

v0.6.13