Bug fix, adds file valid check
Bug fix in plex/library_watchdog for 'file' variable overwrite
when printing log message. This caused processes to hang.
Variable is no longer overwritten.

Vaild file check added to mediainfo class as isValidFile() method.
This method checks if the the file size is greater than the
size of the first video stream. This method is then used in the
DVRconverter class in the convert() method. After plexRename
method is run check to see if file is valid. If valid, then run
comskip and transcode. Else, delete the input file and the
renamed file. Plex Media Scanner is run in all instances because
if file deleted, we want it removed from the library.

The isValidFile check is added because corrupt files that
'successfully' record cause issues in comskip (took all day to
not finish process) and we don't want crappy files hanging around.

To do:
  - Move the check for valid file to be before the plexRename;
    don't need to do all that work if the file is invalid.

v0.6.14