Checks file valid before rename DVR
Completes 'To do' from v0.6.14 and the file is checked for valid
before renaming, comskip, and transcode. This is done to save
time; previous version queried imdb and thetvdb in attempt to get
imdb id before checking file was valid. If file invalid, deleted,
so downloading data to get imdb id was a waste.

Also fixes 'bug' where out_file was not defined in DVRconverter.convert()
method. This was handled fine as convert() method was warpped in
try/except; however, better to reduce exceptions so out_file and
info variables are now defined at start of convert() method to
insure no undefined variable exceptions thrown.

v0.6.15