Updates Plex Media Scanner handling
Running of Plex Media Scanner has been significantly changed.
The plexMediaScanner() function has been added, which takes
action(s) as arguments and items to which actions apply as
keyword arguments (see documentation for Plex Media Scanner).

When the plexMediaScanner module is loaded, the getPlexMediaScanner()
function, modified version of getPlexScannerCMD(), is run to
determine full path to the CLI. We also run the new getPlexLibraries()
function to obtain a dictionary of Plex Libraries; keys are library
names and values are the library section IDs. We can then use this
information (the cmd path and library list) when we run the CLI;
don't search for the CLI and get list of libraries every time we
want to perform a scan.

While the plexDVR_Scan function has been updated to use the new
getPlexMediaScanner() function, plexDVR_Scan is depricated and
will be removed in the future; the MakeMKV_Watchdog still depends
on it for now.

The DVRConverter.convert() method has been updated to use the new
plexMediaScanner function. The function is run after transcode
completes, and then again if original recording is to be removed;
deleting of files no longer handled in the scanner function.

Other changes include removing import of plexDVR_Scan in the
MakeMKV_Watchdog and Plex_DVR_Watchdog CLIs and deleting
the folloing depricated functions from video_utils/plex/utils.py:
  plexDVR_Cleanup
  findFile

v0.7.0