Finishes implementing plexMediaScanner
The MakeMKV_Watchdog class has been updated to use the new
plexMediaScanner function to add new files to Plex library.

Fixes import bug in DVRconverter for the plexMediaScanner function;
the function was not in plex/utils.py, but in its own file.

Updates the plex/library_watchdog recording purge thread. In
previous version, the __purgeRecordings method was run using a
threading.Timer(). This was done so that on a regular interval
recordings would be checked to see if they were more than a day
old. If that old, we assume that the recording failed, so we want
to remove it from the recording list; given a long enough time and
enough failed recordings, the list could become very large.
However, using Timer() lead to some weird code.

In new version, add a __stop Event to the class that is set when
the __run method is finishing; i.e., the watchdog has been stopped.
In the __purgeRecrodings method, we now loop continuously, waiting
on the __stop event, with a timeout of 3 hours, at the beginning
of the while loop. If the wait returns True, then the __stop event
has been set and we return from the method; killing the thread.
Otherwise, lock is obtained and we check recording start times
relative to current time just like in previous version.
Now there are threads being created and destroyed with Timer()
objects.

The Lock attribute of the library_watchdog class has been moved to
'private' attribute.

Some function documention has been updated

v0.7.1