Adds comskip timeout, removes checkIntegrity
After some testing, realized that the ffmpeg checkIntegrity util
did not always identify broken files and, in some cases,
identify files as bad when they seemed fine. So reverted to doing
just the mediainfo check for integrity.

However, a kill() method has been added to subprocManager that
issues a terminate command to all running processes. In the
comskip() method of comremove, now wait 8 hours for process to
finish. If timeout occurs, kill process and continue.

To do:
  - Improve cleanup; if comskip fails, then file is bad so get rid
    of it.

v0.8.4